Move zendesk API closer to where it is used
